streamlit/requirements.txt
streamlit==1.29.0
pandas==2.1.4
psycopg2-binary==2.9.9
plotly==5.18.0
Pillow==10.1.0

